requirements/gpu.txt
# GPU requirements - replaces onnxruntime with onnxruntime-gpu
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
gunicorn>=22.0.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
sqlalchemy[asyncio]>=2.0.35
asyncpg>=0.30.0
pgvector>=0.3.5
alembic>=1.14.0
redis>=5.2.0
insightface>=0.7.3
onnxruntime-gpu>=1.19.0
opencv-python-headless>=4.10.0
numpy>=1.26.0,<2.0.0
python-multipart>=0.0.12
python-dotenv>=1.0.1
httpx>=0.27.0
aiofiles>=24.1.0
